If you paste a URL into the issue field box and the URL matches the URL for an issue in the Jira installation you have configured then it is automagically rewritten to the issue ID (e.g. https://jira.example.com/browse/FOO-1 will become just FOO-1)
Note: Capturing the paste event for a combo box is apparently more hassle than it's worth.  It required going to a lower level than I felt I should.  So, I have used the text changed event instead.  means that if you type the URL manually the same thing happens, but I don't see that as a problem, even if the use case was copy/paste.