Using current screen's working area instead of the main screen's.
Fixes an issue where the window was jumping higher in the screen even
though it was not even close to the bottom of the working area. This
happens when, for example, the main screen has landscape orientation and
the screen where you use the app has portrait orientation.

Also found that comparing to Bottom instead of Height gives more precise
results for some reason. Tested with landscape/portrait monitors, and with
taskbar on top and bottom of screen.