Actual start time (#97)
* Issue #66 Log using actual start time

The start time when logging to Jira was previously just "now minus the time spent"

Now, it records the actual start time (first time you press play on that issue).  When logging we check both this value and the estimate previously used and use the earlier of the two.  This should mean that if you start the timer later log the time spent (with or without a pause/stop in between) then the actual time you started working is used.  However, if you edit the time spent manually to a larger value then the estimate is used.

Note that if you change the issue key without logging the time or resetting, then the start time for the previous issue will still be used.  This is by design - to allow you to correct and mis-typed issue ID.  if you wish to re-set the start time, just click reset.

* Issue #66 Added start time to log time dialogue

Added fields to the log time dialogue to show the start time and allow you to alter it.  Defaults to the value calculated in this issue.